requirements.txt
python-telegram-bot
google-generativeai
nest_asyncio